# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Project {
  id: string;
  name: string;
  /** Short uppercase identifier e.g. "PROJ" */
  key: string;
  description?: string;
  color: string;
  issueCount: number;
  createdAt: string;
}

export interface ProjectStatusInput {
  id?: string;
  name: string;
  color: string;
  position: number;
}

export interface CreateProjectBody {
  name: string;
  key: string;
  description?: string;
  color?: string;
  statuses?: ProjectStatusInput[];
}

export interface UpdateProjectBody {
  name?: string;
  description?: string;
  color?: string;
}

export type ProjectSummaryByStatus = { [key: string]: number };

export type ProjectSummaryByPriority = { [key: string]: number };

export type ProjectSummaryByType = { [key: string]: number };

export type IssuePriority = (typeof IssuePriority)[keyof typeof IssuePriority];

export const IssuePriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export type IssueType = (typeof IssueType)[keyof typeof IssueType];

export const IssueType = {
  task: "task",
  bug: "bug",
  feature: "feature",
  story: "story",
} as const;

export interface Issue {
  id: string;
  projectId: string;
  /** e.g. "PROJ-42" */
  issueKey: string;
  title: string;
  status: string;
  priority: IssuePriority;
  type: IssueType;
  assignee?: string | null;
  reporter: string;
  labels: string[];
  commentCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectSummary {
  projectId: string;
  totalIssues: number;
  byStatus: ProjectSummaryByStatus;
  byPriority: ProjectSummaryByPriority;
  byType: ProjectSummaryByType;
  recentActivity: Issue[];
}

export interface ProjectStatus {
  id: string;
  projectId: string;
  name: string;
  color: string;
  position: number;
  createdAt: string;
}

export interface UpdateProjectStatusesBody {
  statuses: ProjectStatusInput[];
}

export interface Comment {
  id: string;
  issueId: string;
  content: string;
  author: string;
  createdAt: string;
}

export type IssueDetail = Issue & {
  description?: string | null;
  comments: Comment[];
};

export type CreateIssueBodyPriority =
  (typeof CreateIssueBodyPriority)[keyof typeof CreateIssueBodyPriority];

export const CreateIssueBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export type CreateIssueBodyType =
  (typeof CreateIssueBodyType)[keyof typeof CreateIssueBodyType];

export const CreateIssueBodyType = {
  task: "task",
  bug: "bug",
  feature: "feature",
  story: "story",
} as const;

export interface CreateIssueBody {
  title: string;
  description?: string;
  status?: string;
  priority?: CreateIssueBodyPriority;
  type: CreateIssueBodyType;
  assignee?: string;
  reporter?: string;
  labels?: string[];
}

export type UpdateIssueBodyPriority =
  (typeof UpdateIssueBodyPriority)[keyof typeof UpdateIssueBodyPriority];

export const UpdateIssueBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export type UpdateIssueBodyType =
  (typeof UpdateIssueBodyType)[keyof typeof UpdateIssueBodyType];

export const UpdateIssueBodyType = {
  task: "task",
  bug: "bug",
  feature: "feature",
  story: "story",
} as const;

export interface UpdateIssueBody {
  title?: string;
  description?: string;
  status?: string;
  priority?: UpdateIssueBodyPriority;
  type?: UpdateIssueBodyType;
  assignee?: string;
  labels?: string[];
}

export interface CreateCommentBody {
  content: string;
  author: string;
}

export type AttachmentKind =
  (typeof AttachmentKind)[keyof typeof AttachmentKind];

export const AttachmentKind = {
  image: "image",
  text: "text",
} as const;

export interface Attachment {
  id: string;
  issueId: string;
  fileName: string;
  mimeType: string;
  kind: AttachmentKind;
  sizeBytes: number;
  /** Data URL for images, plain text for text files */
  content: string;
  createdAt: string;
}

export type CreateAttachmentBodyKind =
  (typeof CreateAttachmentBodyKind)[keyof typeof CreateAttachmentBodyKind];

export const CreateAttachmentBodyKind = {
  image: "image",
  text: "text",
} as const;

export interface CreateAttachmentBody {
  fileName: string;
  mimeType: string;
  kind: CreateAttachmentBodyKind;
  sizeBytes: number;
  content: string;
}

export type PulseTaskPriority =
  (typeof PulseTaskPriority)[keyof typeof PulseTaskPriority];

export const PulseTaskPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export type PulseTaskStatus =
  (typeof PulseTaskStatus)[keyof typeof PulseTaskStatus];

export const PulseTaskStatus = {
  todo: "todo",
  in_progress: "in_progress",
  in_review: "in_review",
  done: "done",
} as const;

export interface PulseTask {
  issueId: string;
  projectId: string;
  issueKey: string;
  title: string;
  projectName: string;
  priority: PulseTaskPriority;
  status: PulseTaskStatus;
  dueDate?: string | null;
  reason: string;
  order: number;
  estimateBlocks: number;
  score: number;
}

export interface ProjectNextAction {
  projectId: string;
  projectName: string;
  action: string;
  reason: string;
  issueId?: string | null;
  issueKey?: string | null;
  issueTitle?: string | null;
}

export type PulseRiskSeverity =
  (typeof PulseRiskSeverity)[keyof typeof PulseRiskSeverity];

export const PulseRiskSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface PulseRisk {
  type: string;
  severity: PulseRiskSeverity;
  title: string;
  description: string;
  suggestedFix: string;
  projectId?: string | null;
  issueIds: string[];
}

export interface PulseTodayResponse {
  date: string;
  topTasks: PulseTask[];
  projectNextActions: ProjectNextAction[];
  risks: PulseRisk[];
  reviewPrompt: string;
}

export interface PulseRecomputeResponse {
  ok: boolean;
  computedAt: string;
}

export interface FlowSession {
  id: string;
  issueId?: string | null;
  projectId?: string | null;
  startedAt: string;
  endedAt?: string | null;
  note?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface DailyReview {
  id: string;
  date: string;
  summary?: string | null;
  completedIssueIds: string[];
  carriedIssueIds: string[];
  createdAt: string;
  updatedAt: string;
}

export interface StartFlowSessionRequest {
  issueId: string;
}

export interface StartFlowSessionResponse {
  session: FlowSession;
}

export interface StopFlowSessionRequest {
  note?: string;
}

export interface StopFlowSessionResponse {
  session: FlowSession;
}

export interface SaveDailyReviewRequest {
  date: string;
  summary?: string;
  completedIssueIds?: string[];
  carriedIssueIds?: string[];
}

export interface SaveDailyReviewResponse {
  review: DailyReview;
}

export type ListIssuesParams = {
  status?: string;
  priority?: string;
  type?: string;
  assignee?: string;
  search?: string;
};